# OpenAPI.yaml
openapi: 3.0.3
info:
  title: EIAE -- Emissions Ingestion and Analysis Engine
  version: 1.0.0
  description: |
    REST API for a multi-tenant emissions monitoring platform.
    Manages organizations, sites, assets, emission readings, rollup aggregates,
    policies, and violation detection.
servers:
  - url: http://localhost:{port}/v1
    variables:
      port:
        default: "8080"
paths:
...
  /orgs:
    get:
      operationId: listOrganizations
      tags: [Organizations]
      summary: List all organizations
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Organization"
                  total:
                    type: integer
    post:
      operationId: createOrganization
      tags: [Organizations]
      summary: Create an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                description:
                  type: string
                office_location:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /orgs/{org_id}:
    get:
      operationId: getOrganization
      tags: [Organizations]
      summary: Get a single organization
      parameters:
        - $ref: "#/components/parameters/org_id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /orgs/{org_id}/summary:
    get:
      operationId: getOrganizationSummary
      tags: [Organizations]
      summary: Dashboard summary (site count, asset count, last rollup)
      parameters:
        - $ref: "#/components/parameters/org_id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgSummary"

  # ── Sites ──────────────────────────────────────────────────────────────
...
  /asset-types:
    get:
      operationId: listAssetTypes
      tags: [Asset Types]
      summary: List all asset types
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AssetType"
    post:
      operationId: createAssetType
      tags: [Asset Types]
      summary: Create an asset type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetType"

  # ── Assets ─────────────────────────────────────────────────────────────
  /orgs/{org_id}/sites/{site_id}/assets:
    get:
      operationId: listAssets
      tags: [Assets]
      summary: List assets at a site
      parameters:
        - $ref: "#/components/parameters/org_id"
        - $ref: "#/components/parameters/site_id"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Asset"
                  total:
                    type: integer
    post:
      operationId: createAsset
      tags: [Assets]
      summary: Create an asset
      parameters:
        - $ref: "#/components/parameters/org_id"
        - $ref: "#/components/parameters/site_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                asset_type_id:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"

  # ── Recorder Devices ───────────────────────────────────────────────────
  /devices:
    post:
      operationId: createRecorderDevice
      tags: [Recorder Devices]
      summary: Create a recorder device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  enum: [field_device, sensor, satellite]
                org_id:
                  type: string
                  format: uuid
                device_metadata:
                  type: object
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecorderDevice"
  /orgs/{org_id}/devices:
    get:
      operationId: listOrgDevices
      tags: [Recorder Devices]
      summary: List recorder devices for an organization (sensors + satellites)
      parameters:
        - $ref: "#/components/parameters/org_id"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/RecorderDevice"
  /devices/field:
    get:
      operationId: listFieldDevices
      tags: [Recorder Devices]
      summary: List all field devices (not org-scoped)
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/RecorderDevice"

  # ── Readings (Ingest) ─────────────────────────────────────────────────
...
  /orgs/{org_id}/emissions/trend:
    get:
      operationId: getEmissionTrend
      tags: [Emissions]
      summary: Get emission rollup trend for an entity
      parameters:
        - $ref: "#/components/parameters/org_id"
        - name: entity_type
          in: query
          required: true
          schema:
            type: string
            enum: [org, site, asset]
        - name: entity_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: grain
          in: query
          required: true
          schema:
            type: string
            enum: [hour, day, month]
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/EmissionRollup"
  /orgs/{org_id}/emissions/total:
    get:
      operationId: getEmissionTotal
      tags: [Emissions]
      summary: Get lifetime emission total for an entity
      parameters:
        - $ref: "#/components/parameters/org_id"
        - name: entity_type
          in: query
          required: true
          schema:
            type: string
            enum: [org, site, asset]
        - name: entity_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmissionTotalToDate"
  /emissions/refresh:
    post:
      operationId: refreshEmissions
      tags: [Emissions]
      summary: Trigger a manual rollup + violation scan
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                window:
                  type: string
                  default: "48 hours"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string

  # ── Policies ───────────────────────────────────────────────────────────
...
components:
  parameters:
    org_id:
      name: org_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    site_id:
      name: site_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 200
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
  schemas:
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        office_location:
          type: string
          nullable: true
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrgSummary:
      type: object
      properties:
        org_id:
          type: string
          format: uuid
        site_count:
          type: integer
        asset_count:
          type: integer
        last_update:
          type: string
          format: date-time
          nullable: true
...
    AssetType:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Asset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        site_id:
          type: string
          format: uuid
        asset_type_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RecorderDevice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
          nullable: true
        type:
          type: string
          enum: [field_device, sensor, satellite]
        device_metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EmissionRollup:
      type: object
      properties:
        entity_type:
          type: string
          enum: [org, site, asset]
        entity_id:
          type: string
          format: uuid
        bucket_start:
          type: string
          format: date-time
        total_emission:
          type: number
        reading_count:
          type: integer
        min_reading:
          type: number
        max_reading:
          type: number
        unit:
          type: string
        computed_at:
          type: string
          format: date-time
    EmissionTotalToDate:
      type: object
      properties:
        entity_type:
          type: string
        entity_id:
          type: string
          format: uuid
        total_to_date:
          type: number
        reading_count:
          type: integer
        as_of:
          type: string
          format: date-time
...